+++ metadata.rb
@@ -1,7 +1,9 @@
-name             'ganglia'
-maintainer       'YOUR_COMPANY_NAME'
+name             'chef-ganglia'
+maintainer       'Andrew Raymer'
 maintainer_email 'AndrewRaymer@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures ganglia'
+description      'Installs/Configures chef-ganglia'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.6'
+version          '0.2.3'
+
+depends "chef-httpd"